Feat: 회원 탈퇴 서비스에 프로필 이미지 파일 삭제 조건 추가 (#150)

+++ src/main/java/com/umc/finly/domain/member/service/MyPageServiceImpl.java
@@ -171,10 +171,11 @@ public class MyPageServiceImpl implements MyPageService{
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new CustomException(MemberErrorCode.MEMBER_NOT_FOUND));
 
-        // TODO: 1) 프로필 이미지 파일 삭제
-        /*if (member.hasProfileImage()){
+        // 1) 프로필 이미지 파일 삭제
+        if (member.hasProfileImage()){
             imageStorageService.delete(member.getProfileImageUrl());
-        }*/
+            member.clearProfileImage();
+        }
 
         // 2) refresh token 제거
         member.clearRefreshToken();
